Prevent selection of query-index not supported by source
This works by means of a `filterChangeCallback` (see
STSMACOM-47). Since this has to share knowledge with `render` of what
query indexes exist and which are limited to local sources only, the
representation of that information has changed, and moved out of
`render` to the top level.

Fixes UISE-46.

This initial commit includes some rather verbose `console.log`
output. I will remove that in the next commit, but I want it in the
history in case I ever need it back.