Tweak tests
It turns out that "monster comics", from the Inventory module, only
comes up in the "monster" search if mod-ebsco-ekb is not configured
and therefore returning no records. When it _is_ configured, it
returns way more records.

The right fix is to modify the test to click on the "Local" source
facet. But there's no good way to do that right now, since the
checkbox does't have an ID. So rather then wire in a fragile path, I'm
just changing what record the test looks for, for now.

I'll do it right later.