gracefully handle missing filters (#160)
When clicking out of ui-search to a result and then using the back
button to return to search results, the `parseFiltersString` function
would explode because, somehow, it doesn't receive any filters. I dunno
why; but that's a problem for a different day. The fix here is simply to
prevent it from exploding when it doesn't receive any filters.

Refs UISE-110